backend/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
blinker==1.6.2
boto3==1.26.141
botocore==1.29.141
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.9
decorator==4.4.2
exceptiongroup==1.1.2
Flask==2.3.2
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
langchain==0.0.229
langchainplus-sdk==0.0.20
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
moviepy==1.0.3
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
proglog==0.1.10
pydantic==1.10.11
PyJWT==1.7.1
PyPDF2==3.0.1
pytest==7.4.0
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.18
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
whisper==1.1.10
WTForms==3.0.1
yarl==1.9.2
zipp==3.16.2
